Crate oak_tests

Test utilities to help with unit testing of Oak SDK code.

Enums

Profile

Functions

channel_and_interceptor

Build a labeled channel and interceptor suitable for building a client that connects to a Runtime under test.

compile_rust_wasm

Uses cargo to compile a Rust manifest to Wasm bytes.

private_channel_and_interceptor

Build a channel and a label interceptor suitable for building a client that connects to a Runtime under test. The interceptor uses the given key pair to attach a signature and public-key identity tag to every request.

public_channel_and_interceptor

Build a public-untrusted channel and interceptor suitable for building a client that connects to a Runtime under test.

run_single_module

Convenience helper to build and run a single-Node application with the provided entrypoint name.

run_single_module_default

Convenience helper to build and run a single-Node Application using the default name “oak_main” for its entrypoint.

run_single_module_with_config

Convenience helper to build and run a single-Node application with the provided entrypoint name, passing in the provided ConfigMap at start-of-day.

runtime_config

Build the configuration needed to launch a test Runtime instance that runs a single-Node application with the provided entrypoint name.

runtime_config_wasm

Build the configuration needed to launch a test Runtime instance that runs the given collection of Wasm modules, starting with the given module name and entrypoint.