Function oak_tests::start

pub fn start<S: BuildHasher>(
    config: ApplicationConfiguration, 
    entrypoints: HashMap<String, NodeMain, S>
)

Start running the Application under test, with the given initial Node and channel configuration. Because multiple Nodes are linked into the same test, the Nodes should be configured not to define the global extern "C" oak_main(), as this would lead to duplicate symbols. Instead, the entrypoints map should provide a function pointer for each WasmContents entry in the configuration.

Also, Nodes under test should ensure that the oak_log crate does not end being used for logging during tests, either by not calling oak_log::init() at start-of-day (e.g. based on a build configuration), or by calling oak_tests::init_logging() first, and ignoring the subsequent failure from oak_log::init().