Crate oak_tests

Test utilities to help with unit testing of Oak SDK code.

Functions

channel_close
Test-only placeholder for channel creation, which always appears to succeed (but does nothing).

channel_create
Test-only placeholder for channel creation, which always fails.

channel_find
Test-only placeholder for finding a channel by preconfigured port name, which always returns a hard coded value.

channel_read
Test-only implementation fo channel read functionality, which reads a message from the (global) test channel.

channel_write
Test-only implementation of channel write functionality, which writes (only) the data of the provided message to a (global) test channel.

last_message_as_string
Convenience test helper which returns the last message on the global test channel as a string.

reset_channels
Test helper that sets up a new global test channel.

set_read_status
Test helper that injects a failure for future channel read operations.

set_write_status
Test helper that injects a failure for future channel write operations.

wait_on_channels⚠
Test-only implementation of channel wait functionality, which always indicates that all provided channels are ready for reading.