Crate oak_tests

Test utilities to help with unit testing of Oak SDK code.

Structs

WasmEntrypointFullName

Statics

DEFAULT_NODE_NAME

Functions

channel_close
Test version of channel closure.

channel_create⚠
Test version of channel creation.

channel_read⚠
Test implementation of channel read functionality.

channel_write⚠
Test-only implementation of channel write functionality.

grpc_channel_setup_default
Test helper to set up a channel into the (single) Node under test for injected gRPC requests.

init_logging
Test helper to initialize logging via simple_logger.

inject_grpc_request
Test helper to inject a (single) gRPC request message to the Node under test, and collect a (single) response.

last_message_as_string
Convenience test helper which returns the last message on a channel as a string (without removing it from the channel), assuming that only a single Node (with the default internal name) is present and under test.

node_create⚠
Test implementation of dynamic Node creation.

random_get⚠
Test version of random data generation.

reset
Test helper that clears any handle to channel half mappings.

set_read_status
Test helper that injects a failure for future channel read operations.

set_write_status
Test helper that injects a failure for future channel write operations.

start
Start running the Application under test, with the given initial Node and channel configuration. Because multiple Nodes are linked into the same test, the Nodes should be configured not to define the global extern "C" oak_main(), as this would lead to duplicate symbols. Instead, the entrypoints map should provide a function pointer for each WasmConfiguration entry in the configuration.

start_node
Start running a test of a single-Node Application.

stop
Stop the running Application under test.

wait_on_channels⚠
Test implementation of channel wait functionality.