Crate oak_tests

Test utilities to help with unit testing of Oak SDK code.

Functions

compile_rust_wasm
Uses cargo to compile a Rust manifest to Wasm bytes.

grpc_request

run_single_module
Convenience helper to build and run a single-Node application with the given module name, using the provided entrypoint name.

run_single_module_default
Convenience helper to build and run a single-Node Application with the given module name, using the default name "oak_main" for its entrypoint.